# backend/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Trello Clone API
  description: |
    REST API for the Trello clone application built with TypeScript + Express.
    
    ## Resources
    - **Boards** – Top level containers
    - **Lists** – Columns inside a board
    - **Cards** – Tasks inside a list with labels, description and due dates
  version: 1.0.0
  contact:
    name: POC Trello
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /boards:
    get:
      tags: [Boards]
      summary: Get all boards
      operationId: getBoards
      responses:
        '200':
          description: Array of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
    post:
      tags: [Boards]
      summary: Create a new board
      operationId: createBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardDto'
      responses:
        '201':
          description: Board created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '400':
          $ref: '#/components/responses/BadRequest'
  /boards/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Boards]
      summary: Get a board by ID
      operationId: getBoardById
      responses:
        '200':
          description: Board found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Boards]
      summary: Update a board
      operationId: updateBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBoardDto'
      responses:
        '200':
          description: Board updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Boards]
      summary: Delete a board (cascades lists and cards)
      operationId: deleteBoard
      responses:
        '204':
          description: Board deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /boards/{boardId}/lists:
    parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Lists]
      summary: Get all lists for a board
      operationId: getListsByBoard
      responses:
        '200':
          description: Array of lists ordered by position
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardList'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Lists]
      summary: Create a list in a board
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListDto'
      responses:
        '201':
          description: List created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /boards/{boardId}/lists/reorder:
    parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Lists]
      summary: Reorder lists in a board
      operationId: reorderLists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderedIds]
              properties:
                orderedIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Lists reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardList'
  /lists/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    put:
      tags: [Lists]
      summary: Update a list
      operationId: updateList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListDto'
      responses:
        '200':
          description: List updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardList'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Lists]
      summary: Delete a list (cascades cards)
      operationId: deleteList
      responses:
        '204':
          description: List deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /lists/{listId}/cards:
    parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Cards]
      summary: Get all cards in a list
      operationId: getCardsByList
      responses:
        '200':
          description: Array of cards ordered by position
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Cards]
      summary: Create a card in a list
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardDto'
      responses:
        '201':
          description: Card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /lists/{listId}/cards/reorder:
    parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Cards]
      summary: Reorder cards within a list
      operationId: reorderCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderedIds]
              properties:
                orderedIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Cards reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
  /cards/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Cards]
      summary: Get a card by ID
      operationId: getCardById
      responses:
        '200':
          description: Card found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Cards]
      summary: Update a card
      operationId: updateCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardDto'
      responses:
        '200':
          description: Card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Cards]
      summary: Delete a card
      operationId: deleteCard
      responses:
        '204':
          description: Card deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /cards/{id}/move:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [Cards]
      summary: Move a card to another list
      operationId: moveCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listId, position]
              properties:
                listId:
                  type: string
                position:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Card moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    Label:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color code
    Board:
      type: object
      required: [id, title, description, background, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        background:
          type: string
          description: Hex color code for board background
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBoardDto:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          default: ''
        background:
          type: string
          default: '#0052CC'
    UpdateBoardDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        background:
          type: string
    BoardList:
      type: object
      required: [id, boardId, title, position, createdAt, updatedAt]
      properties:
        id:
          type: string
        boardId:
          type: string
        title:
          type: string
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateListDto:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
    UpdateListDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        position:
          type: integer
    Card:
      type: object
      required: [id, listId, boardId, title, description, position, labels, dueDate, createdAt, updatedAt]
      properties:
        id:
          type: string
        listId:
          type: string
        boardId:
          type: string
        title:
          type: string
        description:
          type: string
        position:
          type: integer
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        customFieldValues:
          type: array
          items:
            $ref: '#/components/schemas/CardFieldValue'
    CreateCardDto:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
    UpdateCardDto:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        position:
          type: integer
        listId:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
...
    CardFieldValue:
      type: object
      required: [id, cardId, fieldId, value, createdAt, updatedAt]
      properties:
        id:
          type: string
        cardId:
          type: string
        fieldId:
          type: string
        value:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...